// src/app/models/employee.ts
export default interface Employee {
  EmployeeID: number;
  FirstName: string;
  LastName: string;
  Title: string;
  TitleOfCourtesy: string;
  BirthDate: string;
  HireDate: string;
  Address: string;
  City: string;
  Region: string;
  PostalCode: string;
  Country: string;
  HomePhone: string;
  Extension: string;
  Notes: string;
  ReportsTo: number;
  Salary: number;
}